+++ context/app/static/js/components/Search/SearchWrapper.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import {
   SearchkitManager, SearchkitProvider, SearchBox,
@@ -19,7 +20,8 @@ function DebugItem(props) {
   );
 }
 
-function getByPath(nested, path) {
+function getByPath(nested, field) {
+  const path = field.id;
   let current = nested;
   const pathEls = path.split('.');
   while (pathEls.length) {
@@ -29,6 +31,9 @@ function getByPath(nested, path) {
     } else {
       return null;
     }
+  }
+  if ('translations' in field) {
+    return field.translations[current];
   }
   return current;
 }
@@ -54,7 +59,7 @@ function makeTableComponent(resultFields, detailsUrlPrefix, idField) {
                       href={detailsUrlPrefix + hit._source[idField]}
                       style={{ display: 'block' }}
                     >
-                      {getByPath(hit._source, field.id)}
+                      {getByPath(hit._source, field)}
                     </a>
                   </td>
                 ),
@@ -68,56 +73,49 @@ function makeTableComponent(resultFields, detailsUrlPrefix, idField) {
   };
 }
 
+function MaskedSelectedFilters(props) {
+  const { hiddenFilterIds } = props;
+  const SelectedFilter = (filterProps) => {
+    const style = hiddenFilterIds.indexOf(filterProps.filterId) === -1
+      ? {} : { display: 'None' };
+    // Copy and paste from
+    // http://docs.searchkit.co/v0.8.3/docs/components/navigation/selected-filters.html
+    // plus typo corrections and wrapping div.
+    /* eslint-disable jsx-a11y/click-events-have-key-events */
+    /* eslint-disable jsx-a11y/no-static-element-interactions */
+    return (
+      <div
+        style={style}
+        className={filterProps.bemBlocks.option()
+          .mix(filterProps.bemBlocks.container('item'))
+          .mix(`selected-filter--${filterProps.filterId}`)}
+      >
+        <div
+          className={filterProps.bemBlocks.option('name')}
+        >
+          {filterProps.labelKey}: {filterProps.labelValue}
+        </div>
+        <div
+          className={filterProps.bemBlocks.option('remove-action')}
+          onClick={filterProps.removeFilter}
+        >
+          x
+        </div>
+      </div>
+    );
+    /* eslint-enable */
+  };
+  return <SelectedFilters itemComponent={SelectedFilter} />;
+}
 
-export default function (props) {
+function SearchWrapper(props) {
   const {
-    apiUrl, prefixQueryFields, filters, detailsUrlPrefix,
+    apiUrl, filters, detailsUrlPrefix,
     idField, resultFields, hitsPerPage, debug, httpHeaders,
-    sortOptions = [{
-      label: 'Relevance',
-      field: '_score',
-      order: 'desc',
-      defaultOption: true,
-    }],
-    hiddenFilterIds = [],
-    searchUrlPath = '_search',
+    sortOptions, hiddenFilterIds, searchUrlPath,
   } = props;
   const resultFieldIds = (resultFields.map((field) => field.id)).concat(idField);
   const searchkit = new SearchkitManager(apiUrl, { httpHeaders, searchUrlPath });
-
-  function MaskedSelectedFilters() {
-    const SelectedFilter = (filterProps) => {
-      const style = hiddenFilterIds.indexOf(filterProps.filterId) === -1
-        ? {} : { display: 'None' };
-      // Copy and paste from
-      // http://docs.searchkit.co/v0.8.3/docs/components/navigation/selected-filters.html
-      // plus typo corrections and wrapping div.
-      /* eslint-disable jsx-a11y/click-events-have-key-events */
-      /* eslint-disable jsx-a11y/no-static-element-interactions */
-      return (
-        <div
-          style={style}
-          className={filterProps.bemBlocks.option()
-            .mix(filterProps.bemBlocks.container('item'))
-            .mix(`selected-filter--${filterProps.filterId}`)}
-        >
-          <div
-            className={filterProps.bemBlocks.option('name')}
-          >
-            {filterProps.labelKey}: {filterProps.labelValue}
-          </div>
-          <div
-            className={filterProps.bemBlocks.option('remove-action')}
-            onClick={filterProps.removeFilter}
-          >
-            x
-          </div>
-        </div>
-      );
-      /* eslint-enable */
-    };
-    return <SelectedFilters itemComponent={SelectedFilter} />;
-  }
 
   const filterElements = filters.map((def) => {
     const Filter = filterTypes[def.type];
@@ -139,7 +137,6 @@ export default function (props) {
           <SearchBox
             autofocus
             searchOnChange
-            prefixQueryFields={prefixQueryFields}
           />
           {filterElements}
         </SideBar>
@@ -150,7 +147,7 @@ export default function (props) {
                 'hitstats.results_found': '{hitCount} results found',
               }}
               />
-              <MaskedSelectedFilters />
+              <MaskedSelectedFilters hiddenFilterIds={hiddenFilterIds} />
               <SortingSelector options={sortOptions} />
             </ActionBarRow>
           </ActionBar>
@@ -176,3 +173,54 @@ export default function (props) {
     </SearchkitProvider>
   );
 }
+
+SearchWrapper.propTypes = {
+  apiUrl: PropTypes.string.isRequired,
+  filters: PropTypes.arrayOf(PropTypes.exact({
+    type: PropTypes.oneOf([
+      // Expand as needed; Starting small to catch typos.
+      'RefinementListFilter',
+    ]).isRequired,
+    props: PropTypes.exact({
+      id: PropTypes.string.isRequired,
+      title: PropTypes.string.isRequired,
+      field: PropTypes.string.isRequired,
+      operator: PropTypes.string.isRequired,
+      size: PropTypes.number.isRequired,
+      translations: PropTypes.objectOf(PropTypes.string),
+    }),
+  })).isRequired,
+  detailsUrlPrefix: PropTypes.string.isRequired,
+  idField: PropTypes.string.isRequired,
+  resultFields: PropTypes.arrayOf(PropTypes.exact({
+    id: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    translations: PropTypes.objectOf(PropTypes.string),
+  })).isRequired,
+  hitsPerPage: PropTypes.number.isRequired,
+  debug: PropTypes.bool,
+  httpHeaders: PropTypes.objectOf(PropTypes.string),
+  sortOptions: PropTypes.arrayOf(PropTypes.exact({
+    label: PropTypes.string.isRequired,
+    field: PropTypes.string.isRequired,
+    order: PropTypes.oneOf(['asc', 'desc']).isRequired,
+    defaultOption: PropTypes.bool.isRequired,
+  })),
+  hiddenFilterIds: PropTypes.arrayOf(PropTypes.string),
+  searchUrlPath: PropTypes.string,
+};
+
+SearchWrapper.defaultProps = {
+  debug: false,
+  sortOptions: [{
+    label: 'Relevance',
+    field: '_score',
+    order: 'desc',
+    defaultOption: true,
+  }],
+  hiddenFilterIds: [],
+  searchUrlPath: '_search',
+  httpHeaders: {},
+};
+
+export default SearchWrapper;