+++ context/app/static/js/components/Search/config.js
@@ -1,4 +1,7 @@
-import { filter, field, organFilter, specimenTypeFilter } from './utils';
+import {
+  filter, field,
+  organTranslations, specimenTypeTranslations,
+} from './utils';
 
 
 export const donorConfig = {
@@ -18,16 +21,16 @@ export const donorConfig = {
 
 export const sampleConfig = {
   filters: [
-    organFilter('origin_sample.organ'),
-    specimenTypeFilter('specimen_type'),
+    filter('origin_sample.organ', 'Organ', organTranslations),
+    filter('specimen_type', 'Specimen Type', specimenTypeTranslations),
     filter('donor.group_name', 'Group'),
     filter('created_by_user_displayname', 'Creator'),
   ],
   fields: [
     field('display_doi', 'ID'),
     field('donor.group_name', 'Group'),
-    field('specimen_type', 'Speciment Type'),
-    field('origin_sample.organ', 'Organ'),
+    field('specimen_type', 'Speciment Type', specimenTypeTranslations),
+    field('origin_sample.organ', 'Organ', organTranslations),
   ],
 };
 
@@ -36,15 +39,14 @@ export const datasetConfig = {
     filter('created_by_user_displayname', 'Creator'),
     filter('data_types', 'Data types'),
     filter('donor.group_name', 'Group'),
-    filter('created_by_user_displayname', 'Creator'),
-    specimenTypeFilter('source_sample.specimen_type'),
+    filter('source_sample.specimen_type', 'Specimen Type', specimenTypeTranslations),
     filter('status', 'Status'),
   ],
   fields: [
     field('display_doi', 'ID'),
     field('donor.group_name', 'Group'),
     field('data_types', 'Data Types'),
-    field('origin_sample.organ', 'Organ'),
+    field('origin_sample.organ', 'Organ', organTranslations),
     field('status', 'Status'),
   ]
 };